First iteration of unbooking
Now each participant registers the hash receipt for each transaction.
An unbook operation has beed defined for the coordinator: when called, it will make all participants to a given transaction abort (previous receipt validation with the coordinator, and using the cid to authenticate with the participants).

In case the coordinator is unavailable, another operation (undo) will be defined for the participants, which will require the receipt for validation.