Changed the structure of the client and the interface.
Now the client is a service that runs as long as needed. For test purposes, clientStarter has been added to send a booking request to it (request should theroetically come from the user interface of the application).
This change will allow the transaction coordinator to verify if the client is still alive before committing.